Use C++ standard int types to allow compilation with clang++ also

+++ src/sqlhelper.h
@@ -6,11 +6,11 @@
 namespace hsql {
 
     void printStatementInfo(SQLStatement* stmt);
-    void printSelectStatementInfo(SelectStatement* stmt, uint num_indent);
-    void printImportStatementInfo(ImportStatement* stmt, uint num_indent);
-    void printInsertStatementInfo(InsertStatement* stmt, uint num_indent);
-    void printCreateStatementInfo(CreateStatement* stmt, uint num_indent);
-    void printExpression(Expr* expr, uint num_indent);
+    void printSelectStatementInfo(SelectStatement* stmt, uintmax_t num_indent);
+    void printImportStatementInfo(ImportStatement* stmt, uintmax_t num_indent);
+    void printInsertStatementInfo(InsertStatement* stmt, uintmax_t num_indent);
+    void printCreateStatementInfo(CreateStatement* stmt, uintmax_t num_indent);
+    void printExpression(Expr* expr, uintmax_t num_indent);
 
 } // namespace hsql
